fix for linux

+++ test/test_find.cpp
@@ -11,11 +11,11 @@ int main()
 		DVVector d_values(ctx, "int32_t", 4, h_values);
 		size_t res;
 		TRTC_Find(ctx, d_values, DVInt32(3), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 		TRTC_Find(ctx, d_values, DVInt32(5), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 		TRTC_Find(ctx, d_values, DVInt32(9), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 	}
 
 	{
@@ -23,9 +23,9 @@ int main()
 		DVVector d_values(ctx, "int32_t", 4, h_values);
 		size_t res;
 		TRTC_Find_If(ctx, d_values, Functor(ctx, {}, { "x" }, "        return x>4;\n"), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 		TRTC_Find_If(ctx, d_values, Functor(ctx, {}, { "x" }, "        return x>10;\n"), res);
-		printf("%d\n", res);	
+		printf("%d\n", (int)res);	
 	}
 
 	{
@@ -33,8 +33,8 @@ int main()
 		DVVector d_values(ctx, "int32_t", 4, h_values);
 		size_t res;
 		TRTC_Find_If_Not(ctx, d_values, Functor(ctx, {}, { "x" }, "        return x>4;\n"), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 		TRTC_Find_If_Not(ctx, d_values, Functor(ctx, {}, { "x" }, "        return x>10;\n"), res);
-		printf("%d\n", res);
+		printf("%d\n", (int)res);
 	}
 }
